Add roothide build for Dopamine2-roothide compat

+++ Little16Prefs/RootListController.m
@@ -15,12 +15,22 @@
 }
 
 - (void)respring {
-    NSString *killall = [[NSFileManager defaultManager] fileExistsAtPath:@"/var/jb/usr/bin/killall"]
-        ? @"/var/jb/usr/bin/killall"
-        : @"/usr/bin/killall";
+    NSString *killall = nil;
+    NSArray *candidates = @[@"/var/jb/usr/bin/killall", @"/usr/bin/killall"];
+    for (NSString *path in candidates) {
+        if ([[NSFileManager defaultManager] fileExistsAtPath:path]) {
+            killall = path;
+            break;
+        }
+    }
     pid_t pid;
-    const char *args[] = {[killall UTF8String], "-9", "SpringBoard", NULL};
-    posix_spawn(&pid, args[0], NULL, NULL, (char *const *)args, NULL);
+    if (killall) {
+        const char *args[] = {[killall UTF8String], "-9", "SpringBoard", NULL};
+        posix_spawn(&pid, args[0], NULL, NULL, (char *const *)args, NULL);
+    } else {
+        const char *args[] = {"/bin/launchctl", "kickstart", "-k", "system/com.apple.SpringBoard", NULL};
+        posix_spawn(&pid, args[0], NULL, NULL, (char *const *)args, NULL);
+    }
 }
 
 @end
